Rename networking name to the original convert name
